backend/pyproject.toml
[project]
name = "educomic"
version = "1.0.0"
description = "Educational comic generation platform"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "supabase",
    "openai",
    "httpx",
    "requests",
    "pydantic",
    "pydantic-settings",
    "reportlab",
    "Pillow",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "hypothesis",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
